refactor: split OmniDocBench version adapters
