+++ evalscope/benchmarks/omnidoc_bench/v1_5/omnidoc_bench_adapter.py
