+++ evalscope/benchmarks/omnidoc_bench/v1_5/requirements.txt
