+++ evalscope/benchmarks/omnidoc_bench/v1_6/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) Alibaba, Inc. and its affiliates.
