+++ evalscope/benchmarks/omnidoc_bench/v1_6/omnidoc_bench_v1_6_adapter.py
@@ -18,8 +18,8 @@ from evalscope.api.metric import AggScore, SampleScore, Score
 from evalscope.api.mixin import CodeExecutionSandboxMixin
 from evalscope.api.registry import register_benchmark
 from evalscope.api.sandbox import SandboxEngine, resolve_engine
-from evalscope.benchmarks.omnidoc_bench.omnidoc_bench_adapter import PROMPT_TEMPLATE
 from evalscope.constants import HubType, Tags
+from ..v1_5.omnidoc_bench_adapter import PROMPT_TEMPLATE
 from .sandbox_scorer import PAGE_METRICS, build_scoring_program, parse_scoring_result
 
 # v1.6-only pins: the ModelScope revision and annotation digest identify the supported dataset;
