+++ evalscope/benchmarks/omnidoc_bench/v1_6/sandbox_scorer.py
